1.0.1
Fixed same talent that has multiple iterations in the API quirks and errors. ex: (Wyvern's Claw which has Heavy, Medium and Light variant but it's always the same card the only difference is the requisites)
Added new race values.
Added update reminder of new versions with a label showing which version the current executable is.
Fixed flashing command line when opening the program.